update: namings and refactor
