+++ client/my-sites/plans/jetpack-plans/product-store/context/store-item-info-context.ts
@@ -1,9 +1,9 @@
 import { createContext, useContext } from 'react';
 import { StoreItemInfo } from '../types';
 
-const StoreItemInfoContext = createContext< StoreItemInfo | null >( null );
+const StoreItemInfoContext = createContext< StoreItemInfo >( {} as StoreItemInfo );
 
 export default StoreItemInfoContext;
 
 // A little hack here as we populate the default value in the component it is used.
-export const useStoreItemInfoContext = () => useContext( StoreItemInfoContext ) as StoreItemInfo;
+export const useStoreItemInfoContext = () => useContext( StoreItemInfoContext );
