+++ client/my-sites/plans/jetpack-plans/product-store/index.tsx
@@ -2,7 +2,7 @@ import { useState } from '@wordpress/element';
 import { useSelector } from 'react-redux';
 import StoreFooter from 'calypso/jetpack-connect/store-footer';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
-import StoreItemInfoContext from './context/store-item-context';
+import StoreItemInfoContext from './context/store-item-info-context';
 import { useStoreItemInfo } from './hooks/use-store-item-info';
 import { ItemsList } from './items-list';
 import { JetpackFree } from './jetpack-free';
